Add lp.setupi instruction
Hardware loops now have their own adder and no longer share it with the jump
target calculation.

The new lp.setupi instruction makes it possible to truly setup hardware loops
with a single instruction. For the lp.setup instruction, a register with the
counter had to be prepared first. The range of the new instruction is quite
limited though, it uses the shifted z-imm (5 bit, unsigned).