Add new testbench for RI5CY
This testbench currently runs RI5CY in a stand-alone fashion. tb_top.sv is the
toplevel module that instantiates the core and a memory mapping module that
contains a RAM model (reused from the verilator testbench) and a pseudo
peripheral, which just dumps written data to stdout.

Currently two types of programs can be run: a small internal program that just
loops infinitely and does some I/O when passing no +firmware= parameter or one
that makes use of the firmware parameter by loading the given program, which
must be a file that on each line has a byte in hexadecimal notation, verbatim
with readmemh into the RAM model.

There is a firmware forked from picorv in the picorv_firmware folder, but as of
now it doesn't work properly yet since our tb is missing a timer (for
interrupts) and uses a different interrupt interface.

Call `make tb-run` for building the testbench and running the small internal
program.

Call `make firmware-run` to build the firmware and testbench, and run it (broken).

Plusarg options for the simulator:
+firmware=  firmware path
+vcd  	    make a vcd, written into riscy_tb.vcd
+verbose    write more information to stdout