Add single stepping mode
rtl/riscv_controller:
We enter re-enter debug mode immediately again after executing one instruction
when dsingle_step_i is one. This is currently done by checking said flag during
the DECODE step. To prevent taking this if-branch when we are in the debug ROM
we have to make sure we are not in debug_mode. Normally we enter DBG_FLUSH,
DBG_TAKEN_IF (saves dpc) and go to the debug ROM aftwards.

There are some complications with branches: When we have a branch instruction in
the ID stage we can't provide a value to the dpc register yet (which needs to be
the next to be executed instruction) instead we enter an immediate state called
DBG_WAIT_BRANCH where we process the branch taken/not taken decision from the EX
stage and then proceed to DBG_FLUSH and DBG_TAKEN_IF.

rtl/riscv_cs_registers:
Consider debug_mode_i before saving to mstatus and mepc. In debug mode we don't
save regulary, instead we use a separate register (dcsr.prv) to save and restore
the privilege mode. This makes sense since we only want to observe the hart not
disturb it by making it go through the privilege change motions.

rtl/riscv_core:
rtl/riscv_id_stage:
Route new signals.