Fix single stepping over ecall and xret
We need to execute more of the regular controller for ecall and xrets to
properly grab the target address to store into dpc. This is done
similarily as for the illegal instruction: we go through FLUSH_WB as would
be done when we weren't single stepping and fork there by checking for
the single stepping condition.

* Rewrite single step handling in DECODE state

* Add single stepping over ebreak when ebreakm is low